+++ lib/features/home/presentation/all_friends_page.dart
@@ -84,7 +84,7 @@ class _FriendGridItem extends StatelessWidget {
   final String friendId;
   final String name;
   final int level;
-  final int faithLevel;
+  final double faithLevel;
 
   const _FriendGridItem({
     required this.friendId,
