+++ lib/features/home/providers/faith_history_recorder.dart
@@ -1,6 +1,6 @@
 import 'package:supabase_flutter/supabase_flutter.dart';
 
-Future<void> recordFaithSnapshot(String userId, int faithLevel) async {
+Future<void> recordFaithSnapshot(String userId, double faithLevel) async {
   final today = DateTime.now().toIso8601String().substring(0, 10);
   await Supabase.instance.client.from('faith_history').upsert(
     {
